Fix policies issues * Add seed for user with role user * Isolate API Auth logic to controller * Start using policies on API controller

+++ src/Http/Controllers/BaseAPIController.php
@@ -19,8 +19,7 @@ class BaseAPIController extends Controller
 
     public function index()
     {
-        // TODO
-        // $this->authorize('viewAny', $this->resource['model_fqn']);
+        $this->authorize('viewAny', $this->resource['model_fqn']);
 
         return $this->resource['api_resource_fqn']::collection(
             $this->resource['model_fqn']::latest()
@@ -32,16 +31,14 @@ class BaseAPIController extends Controller
     {
         $model = $this->resource['model_fqn']::findByRouteKeyName($model)->firstOrFail();
 
-        // TODO
-        // $this->authorize('update', $model);
+        $this->authorize('view', $model);
 
         return new $this->resource['api_resource_fqn']($model);
     }
 
     public function store()
     {
-        // TODO
-        // $this->authorize('create', $this->resource['model_fqn']);
+        $this->authorize('create', $this->resource['model_fqn']);
 
         $request = app($this->resource['request_fqn']);
 
@@ -54,10 +51,9 @@ class BaseAPIController extends Controller
 
     public function update(Request $request, $model)
     {
-        // TODO
-        // $this->authorize('update', $this->resource['model_fqn']::findByRouteKeyName($model)->firstOrFail());
-
         [$request, $model] = $this->bindModelAndValidateRequest($request, $model);
+
+        $this->authorize('update', $model);
 
         $model->updateWithRelations($request);
 
@@ -70,8 +66,7 @@ class BaseAPIController extends Controller
     {
         $model = $this->resource['model_fqn']::findByRouteKeyName($model);
 
-        // TODO
-        // $this->authorize('delete', $model->firstOrFail());
+        $this->authorize('delete', $model->firstOrFail());
 
         $model->delete();
